% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batdataset.R
\docType{data}
\name{bat}
\alias{bat}
\title{Bat datset}
\format{A list of two items
\describe{
  \item{\code{tree}}{the phylogenetic tree object}
  \item{\code{traitset}}{a data frame of trait data}
}}
\usage{
data(bat)
}
\description{
A dataset containing a phylogenetic tree and trait data for bat species
}
\references{
Aguirre, Luis F., et al. "Ecomorphological analysis of trophic niche partitioning in a tropical savannah bat community." Proceedings of the Royal Society of London. Series B: Biological Sciences 269.1497 (2002): 1271-1278.
}
\keyword{datasets}
