% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coraldataset.R
\docType{data}
\name{coral}
\alias{coral}
\title{Coral datset}
\format{A list of two items
\describe{
  \item{\code{tree}}{the phylogenetic tree object}
  \item{\code{traitset}}{a data frame of trait data}
}}
\usage{
data(coral)
}
\description{
A dataset containing a phylogenetic tree and trait data for coral species
}
\references{
Sanchez, Juan Armando, and Howard R. Lasker. "Patterns of morphological integration in marine modular organisms: supra-module organization in branching octocoral colonies." Proceedings of the Royal Society of London. Series B: Biological Sciences 270.1528 (2003): 2039-2044.
}
\keyword{datasets}
