% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lizarddataset.R
\docType{data}
\name{lizard}
\alias{lizard}
\title{Lizard datset}
\format{A list of two items
\describe{
  \item{\code{tree}}{the phylogenetic tree object}
  \item{\code{traitset}}{a data frame of trait data}
}}
\usage{
data(lizard)
}
\description{
A dataset containing a phylogenetic tree and trait data for lizard species
}
\references{
Molina-Borja, M., & Rodriguez-Dominguez, M. A. (2004). Evolution of biometric and life-history traits in lizards (Gallotia) from the Canary Islands. Journal of Zoological Systematics and Evolutionary Research, 42(1), 44-53.
}
\keyword{datasets}
