% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ouxy.r
\name{oubmcirmodel}
\alias{oubmcirmodel}
\title{Simulate traits under OUBMCIR model given a set of parameters}
\usage{
oubmcirmodel(model.params, reg.params, root = root, tree = tree)
}
\arguments{
\item{model.params}{A vector of model parameters including alpha.y: force parameter of response trait, sigmasq.x: rate parameter of covariate, alpha.tau: force parameter of rate, theta.tau optimum parameter of rate, sigmasq.tau: rate parameter of rate}

\item{reg.params}{A vector of regression paramters including  b0: intercept parameter of regression, b1: slope parameter of first covariate,  b2: slope paramter of the second covariate}

\item{root}{A vector of numerical values for root of species estimated from \code{OUprior}}

\item{tree}{An ape: tree object stored in phylo format}
}
\value{
Returns the trait vectors \eqn{Y=(y_1,y_2,\cdots,y_n)', X_1=(x_{1,1},x_{1.2},\cdots,x_{1,n})',X_2=(x_{2,1},x_{2,2},\cdots,x_{2,n})'} simulated from the model.
}
\description{
Simulate traits under OUBMCIR model given a set of model parameters, regression parameters, tree and ancestral values.
}
\details{
The model requires user to input  model parameters \eqn{\alpha_y,\sigma^2_x,\alpha_\tau,\theta_\tau,\sigma^2_\tau} and regression parameters \eqn{b_0,b_1,b_2}, a tree object, trait dataset (one response, two covariate), ancestral values(root) which is estimated by BM or OU model from \code{\link[geiger:fitContinuous]{geiger}}. Then the algorithm starts from the root and apply the tree traversal algorithm to simulate trait on each node according the OUOUBM dynamics.
}
\examples{
library(ape)
tree<-rcoal(5)
tree<-reorder(tree,"postorder")
root<-list(y.ou.sigmsq=1 ,y.ou.root=0, x1.ou.root=0, x2.ou.root=0,x1.bm.root=0, x2.bm.root=0)
model.params<-c(0.5,0.25,0.3,1,0.2)
names(model.params)<-c("alpha.y","alpha.x","theta.x","sigmasq.x","sigmasq.tau")
reg.params<-c(0,1,1)
names(reg.params)<-c("b0","b1","b2")
oubmcirmodel(model.params,reg.params,root=root,tree=tree)

}
\references{
\enumerate{
  \item Jhwueng, D-C. (2019) Statistical modeling for adaptive trait evolution. Under review.
  \item Jhwueng, D-C., and Vasileios Maroulas. "Adaptive trait evolution in random environment." Journal of Applied Statistics 43.12 (2016): 2310-2324.
  \item Hansen, Thomas F., Jason Pienaar, and Steven Hecht Orzack. "A comparative method for studying adaptation to a randomly evolving environment." Evolution: International Journal of Organic Evolution 62.8 (2008): 1965-1977.
}
}
